+++ parser_library/src/diagnostic.h
@@ -825,11 +825,9 @@ public:
 
     static diagnostic_s error_W0004(std::string_view file_name, std::string_view ws_name);
 
-    static diagnostic_s error_W0005(std::string_view file_name, std::string_view proc_group);
+    static diagnostic_s error_W0005(std::string_view file_name, std::string_view name, std::string_view type);
 
     static diagnostic_s error_W0006(std::string_view file_name, std::string_view proc_group);
-
-    static diagnostic_s error_W0007(std::string_view file_name, std::string_view proc_group);
 
     /*
     E01x - wrong format