+++ parser_library/src/workspaces/processor_group.cpp
@@ -14,8 +14,6 @@
 
 #include "processor_group.h"
 
-#include "config/proc_grps.h"
-
 namespace hlasm_plugin::parser_library::workspaces {
 
 namespace {
@@ -33,47 +31,14 @@ struct translate_pp_options
 };
 } // namespace
 
-processor_group::processor_group(const std::string& pg_name,
-    std::string_view pg_file_name,
-    const config::assembler_options& asm_options,
-    const config::preprocessor_options& pp)
+processor_group::processor_group(
+    const std::string& pg_name, const config::assembler_options& asm_options, const config::preprocessor_options& pp)
     : m_pg_name(pg_name)
-    , m_asm_opts(translate_assembler_options(asm_options, pg_file_name))
+    , m_asm_opts(asm_options)
     , m_prep_opts(std::visit(translate_pp_options {}, pp.options))
 {}
 
-instruction_set_version processor_group::find_instruction_set(
-    std::string_view optable, std::string_view pg_file_name) const
-{
-#ifdef __cpp_lib_ranges
-    auto it = std::ranges::lower_bound(
-        instr_set_version_equivalents, optable, {}, [](const auto& instr) { return instr.first; });
-#else
-    auto it = std::lower_bound(std::begin(instr_set_version_equivalents),
-        std::end(instr_set_version_equivalents),
-        optable,
-        [](const auto& l, const auto& r) { return l.first < r; });
-#endif
-
-    if (it == std::end(instr_set_version_equivalents) || it->first != optable)
-    {
-        add_diagnostic(diagnostic_s::error_W0007(pg_file_name, m_pg_name));
-        return asm_option::instr_set_default;
-    }
-
-    return it->second;
-}
-
-asm_option processor_group::translate_assembler_options(
-    const config::assembler_options& asm_options, std::string_view pg_file_name) const
-{
-    return asm_option { .sysparm = asm_options.sysparm,
-        .profile = asm_options.profile,
-        .instr_set = asm_options.optable.empty() ? asm_option::instr_set_default
-                                                 : find_instruction_set(asm_options.optable, pg_file_name),
-        .system_id = asm_options.system_id.empty() ? asm_option::system_id_default : asm_options.system_id,
-        .sysopt_xobject = asm_options.goff };
-}
+void processor_group::update_asm_options(asm_option& opts) const { m_asm_opts.apply(opts); }
 
 void processor_group::collect_diags() const
 {