+++ clients/vscode-hlasmplugin/CHANGELOG.md
@@ -9,6 +9,7 @@
 - Instruction set versioning support
 - Basic GOFF, XOBJECT and SYSOPT_XOBJECT support
 - MNOTE support
+- Assembler options can be specified in pgm_conf.json to override values from proc_grps.json
 
 #### Fixed
 - Fixed an issue preventing correct N' attribute evaluation of empty subscript arrays